mysite/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
sqlparse==0.4.4
